requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
backports.zoneinfo==0.2.1
billiard==4.1.0
build==0.10.0
celery==5.3.1
certifi==2023.7.22
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.1
email-validator==2.0.0.post2
environs==9.5.0
exceptiongroup==1.1.2
fastapi==0.100.1
flower==2.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
marshmallow==3.20.1
orjson==3.9.2
packaging==23.1
pip-tools==7.1.0
pluggy==1.2.0
pre_commit==3.3.2
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
pyproject_hooks==1.0.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
responses==0.23.2
sentry-sdk==1.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tomli==2.0.1
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
